backend/django/pyproject.toml
[tool.poetry]
name = "devmind-django"
version = "0.1.0"
description = "DevMind Django backend"
authors = ["Muhammed Sinan <muhammedsiinanmp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3"
django-environ = "^0.11"
celery = {extras = ["redis"], version = "^5.4"}
django-celery-beat = "^2.6"
django-celery-results = "^2.5"
psycopg = {extras = ["binary"], version = "^3.1"}
dj-database-url = "^2.1"
httpx = "^0.27"
sentry-sdk = {extras = ["django"], version = "^2.5"}
cryptography = "^42.0"
PyGithub = "^2.3"
redis = "^5.0"
django-redis = "^5.4"
whitenoise = "^6.6"
gunicorn = "^22.0"
requests = "^2.33.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-django = "^4.8"
pytest-cov = "^5.0"
factory-boy = "^3.3"
responses = "^0.25"
freezegun = "^1.4"
mypy = "^1.10"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0"}
djangorestframework-stubs = "^3.15"
ruff = "^0.4"
black = "^24.4"
django-extensions = "^3.2"
types-requests = "^2.33.0.20260408"
bandit = "^1.9.4"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "devmind.settings.test"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=apps --cov-report=term-missing --cov-fail-under=95"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "devmind.settings.base"

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.black]
line-length = 88
target-version = ["py312"]
